Restrict picks on games that already began
In addition:
Only display current users picks

+++ app/controllers/picks_controller.rb
@@ -1,25 +1,60 @@
 class PicksController < ApplicationController
   def set_game_pick
-    selected_team = params[:selected_team]
-    game_id = params[:game_matchup_id]
 
-    pick = Pick.find_by(user_id: current_user.id, game_id: game_id)
+    if params[:time_now] < params[:game_start_time]
+      #Game Not Started
 
-    if pick.nil?
-      Pick.create(user_id: current_user.id, game_id: game_id, picked_team_id: selected_team)
+      selected_team = params[:selected_team]
+      game_id = params[:game_matchup_id]
+
+      pick = Pick.find_by(user_id: current_user.id, game_id: game_id)
+
+      if pick.nil?
+        Pick.create(user_id: current_user.id, game_id: game_id, picked_team_id: selected_team)
+      else
+        pick.update_attribute("picked_team_id", selected_team)
+      end
+
+      render json: { game: game_id, team: selected_team }
+
     else
-      pick.update_attribute("picked_team_id", selected_team)
-    end
+      #Game Already started
 
-    render json: { game: game_id, team: selected_team }
+    end  
+
+    # selected_team = params[:selected_team]
+    # game_id = params[:game_matchup_id]
+
+    # pick = Pick.find_by(user_id: current_user.id, game_id: game_id)
+
+    # if pick.nil?
+    #   Pick.create(user_id: current_user.id, game_id: game_id, picked_team_id: selected_team)
+    # else
+    #   pick.update_attribute("picked_team_id", selected_team)
+    # end
+
+    # render json: { game: game_id, team: selected_team }
   end
 
   def clear_pick
-    game = params[:game_matchup_id]
+    if params[:time_now] < params[:game_start_time]
+      #Game Not Started
+      game = params[:game_matchup_id]
 
-    pick = Pick.find_by(game_id: game, user_id: current_user.id)
-    pick.destroy
+      pick = Pick.find_by(game_id: game, user_id: current_user.id)
+      pick.destroy
 
-    render json: { game: game }
+      render json: { game: game }
+    else
+      #Game Already started
+    end
+
+
+    # game = params[:game_matchup_id]
+
+    # pick = Pick.find_by(game_id: game, user_id: current_user.id)
+    # pick.destroy
+
+    # render json: { game: game }
   end
 end
